feat: rm HashedStateChanges, introduce StorageWriter::write_hashed_state
This phases out the `HashedStateChanges` type, instead using the
`StorageWriter` struct for writing sorted hashed state.
