+++ crates/storage/provider/src/bundle_state/mod.rs
@@ -2,11 +2,9 @@
 //! This module contains all the logic related to bundle state.
 
 mod execution_outcome;
-mod hashed_state_changes;
 mod state_changes;
 mod state_reverts;
 
 pub use execution_outcome::{AccountRevertInit, BundleStateInit, OriginalValuesKnown, RevertsInit};
-pub use hashed_state_changes::HashedStateChanges;
 pub use state_changes::StateChanges;
 pub use state_reverts::{StateReverts, StorageRevertsIter};
